+++ src/config.default.js
@@ -34,6 +34,12 @@ module.exports = {
     },
   ],
 
+  // Custom Playwright options: https://playwright.dev/docs/api/class-browser#browser-new-context
+  playwrightOptions: {
+    // example: use dark mode
+    // colorScheme: "dark",
+  },
+
   // Handles HTTP status codes above 400 as errors, i.e., evaluatory modules will not be executed.
   handleHttpErrorCodes: true,