+++ src/utils/__tests__/config.test.js
@@ -11,13 +11,13 @@ describe("config", () => {
   });
 
   test("returns the default config if no config file is specified", async () => {
-    expect(await getConfig()).toEqual(defaultConfig);
+    await expect(getConfig()).resolves.toEqual(defaultConfig);
   });
 
   test("user config overrides default config", async () => {
-    expect(
-      await getConfig({ config: join(__dirname, "test-config.json") })
-    ).toEqual({
+    await expect(
+      getConfig({ config: join(__dirname, "test-config.json") })
+    ).resolves.toEqual({
       ...defaultConfig,
       maxConnections: 1,
       modules: ["pa11y"],
@@ -25,12 +25,12 @@ describe("config", () => {
   });
 
   test("cli argument overrides user config", async () => {
-    expect(
-      await getConfig({
+    await expect(
+      getConfig({
         config: join(__dirname, "test-config.json"),
         modules: "axe-core",
       })
-    ).toEqual({
+    ).resolves.toEqual({
       ...defaultConfig,
       maxConnections: 1,
       modules: ["axe-core"],
@@ -39,10 +39,10 @@ describe("config", () => {
 
   test("returns the default config if the specified config file doesn't exist", async () => {
     logger.setLevel("silent");
-    expect(
-      await getConfig({
+    await expect(
+      getConfig({
         config: join(__dirname, "nonexisting-config.json"),
       })
-    ).toEqual(defaultConfig);
+    ).resolves.toEqual(defaultConfig);
   });
 });